allow disabling processing for formatters and index names
The formatters and elasticsearch_index_name configurations now have
an `enabled` parameter.  This defaults to `true`.
This is useful for when you want to have a single plugin
configuration shared among multiple fluentd used as both collectors
and normalizers, and you don't want the collectors to do the
elasticsearch index name creation because they aren't talking directly
to elasticsearch.