Bug 1514110 - Aggregated Logging replacing all log levels with '3' and '6' after upgrade to 3.5 from 3.4
https://bugzilla.redhat.com/show_bug.cgi?id=1514110
If there is already a `level` field in the record, then see if it is
a "close" match to one of the canonical `level` field values at
https://github.com/ViaQ/elasticsearch-templates/blob/master/namespaces/_default_.yml#L63
e.g. if `level` is "CRITICAL", convert to "crit", if level
is "WARN" convert to 'warning', etc.

Otherwise, if we cannot use it directly or normalize it, convert
it to its string representation (ruby `to_s` method) and store
the string value in the `level` field.

This commit cleans up other `level` processing and adds several tests
for `level` processing.