Bug 1756920: fluentd pods do not process kubernetes events
https://bugzilla.redhat.com/show_bug.cgi?id=1756920
Add a per-formatter `process_kubernetes_events`.  This will
allow us to always have event processing enabled for eventrouter
logs, but disabled for everything else, assuming the user creates
eventrouter logs with a specific name, as specified in the docs.
For example, if the eventrouter is deployed with a pod name matching
`eventrouter-*`, we can do something like this:

```
    <filter **>
      @type viaq_data_model
      ...
      process_kubernetes_events false
      <formatter>
        tag "kubernetes.var.log.containers.eventrouter-**"
        type k8s_json_file
        remove_keys log,stream,CONTAINER_ID_FULL,CONTAINER_NAME
        process_kubernetes_events true
      </formatter>
      <formatter>
        tag "kubernetes.var.log.containers**"
        type k8s_json_file
        remove_keys log,stream,CONTAINER_ID_FULL,CONTAINER_NAME
      </formatter>
```

Then, we can have eventrouter process on by default, always, with no
env. vars or api changes required.